Add module options
Might need to clarify these a bit still. The first option determines whether or not the
gallery page is accessible to logged-out users. The second option determines whether
we show pages from all institutions, from no institutions (which disables the page),
or from just some institutions. And the last option is which institutions to use if
you've selected some institutions.